Reorganizacion de carpetas e instrucciones de instalacion del FX
