+++ .idea/libraries/JavaFX_Library.xml
@@ -1,0 +1,10 @@
+<component name="libraryTable">
+  <library name="JavaFX_Library">
+    <CLASSES>
+      <root url="file://$USER_HOME$/Documents/myJavaStuff/javafx-sdk-18.0.1/lib" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+    <jarDirectory url="file://$USER_HOME$/Documents/myJavaStuff/javafx-sdk-18.0.1/lib" recursive="false" />
+  </library>
+</component>
