+++ Importante/path.txt
@@ -1,0 +1,10 @@
+El siguiente es una dirección o path para poder abrir las clases que sean aplicaciones de JavaFX:
+
+--module-path "\path\to\javafx-sdk-11\lib" --add-modules javafx.controls,javafx.fxml
+
+
+En mi caso, el path de mi pc es el sigueinte:
+--module-path "C:\Users\USUARIO\Documents\myJavaStuff\javafx-sdk-18.0.1\lib" --add-modules javafx.controls,javafx.fxml
+
+
+
