+++ src/application/application.java
@@ -1,0 +1,16 @@
+package application;
+
+import javafx.application.Application;
+import javafx.stage.Stage;
+
+public class application extends Application {
+
+    public static void main(String[] args) {
+        launch(args);
+    }
+
+    @Override
+    public void start(Stage primaryStage) {
+
+    }
+}
