+++ src/model/Tienda.java
@@ -1,4 +1,4 @@
-package Aplicacion;
+package model;
 
 public class Tienda {
